+++ packages/vendure-plugin-store-credit/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0 (2026-02-03)
+
+- Accept amount as input to the store credit payment handler
+- Export helper functions to populate wallets for customers
+
 # 1.0.0 (2026-02-03)
 
 - Initial release